refactor: create execution/headless/ subpackage
Move headless.py → headless/__init__.py and 5 _headless_*.py modules
into execution/headless/. Update all internal cross-references.
